refactor(aisections): extract useAISectionsBulkTransform hook
Phase 3 of the V12 overlay split — the entire transform state machine
collapses into one hook. The overlay body now renders JSX against the
hook's return value plus a thin selection-pick layer that needs onSelect
(which the hook doesn't get to keep its dependency surface tight).

The hook owns:
- drag state + snap/cascade toggles
- bulk flattening (`bulkRefs`, `bulkEntityCount`, `isBulkActive`,
  `bulkSectionIndices`)
- pivot snapshot + override + dragging state, with membership-keyed reset
- gizmo target selection (bulk > single-entity per ADR-0010)
- preview model / section / corners / cascade-neighbours
- gizmo position + axes + pixelSize
- gesture handlers (`handleGizmoTransform/Commit/Cancel`)
- pivot drag handlers (`handlePivotMove/Commit/Cancel`)
- numeric-panel session pivot + handlers (issue #81)
- transient edge / drag UI reset on section change

Session publish to `BulkTransformGizmoSessionProvider` is the one
legitimate effect — extracted into the file-local `useGizmoSessionPublish`
sub-hook per CLAUDE.md "Don't reach for useEffect", with the on-unmount
clear in its own effect inside the same hook.

AISectionsOverlay.tsx: 1172 → 554 lines (target ≤500, slightly over).
useAISectionsBulkTransform.ts: 714 lines (the user accepted "one fat
hook" over splitting; flagging for re-discuss per Definition of done).

Refs #48, #59 (partial — WorkspaceHierarchy still deferred).
